requirements.txt
piexif==1.1.3
Pillow==8.3.1